Push PKGBUILD update after repo deploy
This lets failed deploys be retried without waiting for a new version to be released or manually retrying the failed job. The original choice of leaving the deploy until last was intentional, but due to the increased use of the Pacman repo, which is likely higher than people building the package themselves, it makes sense to prioritise the deploy being up-to-date over the repo's PKGBUILD. In the future, the deploy could compare with the last deployed version instead of the repo's PKGBUILD, but if GitHub's failure rate is low enough, this should keep the repo up-to-date enough.